+++ Bioteca.Prism.Core/Security/User.cs
@@ -4,5 +4,6 @@ public class User
 {
     public Guid Id { get; set; }
     public string Login { get; set; }
+    public Guid? ResearcherId { get; set; }
 }